Correct Adwords to AdWords everywhere.
For members of the public API the old names have been left around as
aliases, so there should not be any compatibility problems. Unless you
were doing things like hardcoding the name of an exception as a string.
Which is, I suppose, quite feasible. Still, I’m not going to consider it
a breaking change.