v201603 update: includeZeroImpressions change
Formerly it was a part of the report definition; this is finally removed
in favour of the HTTP header approach.

This is a breaking change. If you had `'includeZeroImpressions'` in your
report definition for a `ReportFile.objects.request()` call you will
need to remove it from the report definition and start passing it as a
keyword argument named `include_zero_impressions` instead.